Document groups in AttributeProvider::group_end
I saw the TODO comment here, and was also looking for this information
myself. So I ended up reading up on them in the BLE spec, and this is
what I came up with! I'm about 75% confident in my definitions here -
it accurately captures what I understand, but I haven't found
everything related to grouping so there's a possibility this is wrong
and/or not what was intended by the TODO.

Signed-off-by: David Ross <David.Ross@wdc.com>