Use a trait for representing attribute values
This makes it easier to use attributes in environments where using
static lifetimes for attribute values is challenging by allowing the
attribute service to use its own types that can choose how to store the attribute
values.

In addition, the usage of attribute in the AttributeProvider now takes a
reference to the Attribute instead of a copy to allow the attribute
service more freedom in implementing.